Update changelog, move tests
